Adds a role to someone from a set list.
Adds a role to member, if they are listening to spotify (which counts as a game) this will be ignored.

A dictionary list can be added to by anyone in the server, the bot will search this list for the correct role to add to someone, for example the role 'gta' may be an abreviation of the game 'Grand Theft Auto V'.