Role given to member, created if does not exist.
When a member updates their presence, the bot will check to see if the new presence is a whitelisted application, these applications can also have abreviated role names, for example. In the program currently, the bot will check to see if someone is playing League of Legends, and if they are, assign that person the 'league' role, if it exists, if the role does not exist, the bot will create it, then give the person the role.